requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
Django==3.2.8
django-filter==21.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
docxcompose==1.3.3
docxtpl==0.14.2
Jinja2==3.0.2
lxml==4.6.3
MarkupSafe==2.0.1
mysqlclient==2.0.3
Pillow==8.4.0
PyJWT==2.3.0
pypandoc==1.6.4
python-docx==0.8.11
python-dotenv==0.19.1
pytz==2021.3
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.2
uWSGI==2.0.20
xlwt==1.3.0